Added best model for mask detection(best_mobilenet_mask_detector.pt) and updated training script with out vusilization.
The train.py script has been updated to include the best model for mask detection, which is now saved as `best_mobilenet_mask_detector.pt`. This model is trained using the MobileNet architecture and is intended for detecting whether individuals are wearing masks.
Also excludes vusilization and unnecessary dependencies from the training process to streamline the workflow.